// src/app/model/test3.ts
export interface test3 {
    id4: number;
    inputDate: string;
    inputDay: string;
    inputTime: string;
    inputLocation: string;
    inputP: string;
    inputS: string;
    inputZn: string;
    inputFe: string;
    inputMn: string;
    inputCu: string;
    inputK: string;
    inputCa: string;
    inputMg: string;
    inputNa: string;
    inputOM: string;
    inputCEC: string;
    inputpH: string;
}